Implement eightball rule type with shared PoolRules base
- Introduced `PoolRules` abstract class to consolidate logic common to pool games (NineBall, EightBall).
- Implemented EightBall rules with autonomous class implementing `Rules` directly.
- Added requested Eightball rack arrangement to `Rack`.
- Synchronized group assignment using `p1type` field in `ScoreEvent`.
- Updated HUD to reflect assigned groups.
- Fixed SonarCloud duplication issues by refactoring NineBall and EightBall to extend PoolRules.
- Updated `Rules` interface to handle score synchronization centrally.
