+++ src/controller/controllerbase.ts
@@ -36,7 +36,7 @@ export abstract class ControllerBase extends Controller {
   }
 
   override handleScore(event: ScoreEvent): Controller {
-    this.container.updateScoreHud(event.p1, event.p2, event.b, event.active)
+    this.container.rules.handleScore(event)
     return this
   }
